@@ -29,37 +29,54 @@
 	],
 	"main": "dist/index.js",
 	"types": "dist/index.d.ts",
+	"type": "module",
 	"files": [
 		"dist/"
 	],
 	"scripts": {
+		"clean": "del-cli dist coverage",
 		"build": "tsc",
-		"pretest": "npm run build",
-		"lint": "eslint .",
-		"autofix": "npm run lint -- --fix",
-		"test": "npm run lint && cross-env TS_NODE_PROJECT=test/tsconfig.json mocha --require ts-node/register test/*-test.*",
-		"precommit": "pretty-quick --staged",
-		"prepublish": "npm run build"
+		"compile": "run-s clean build",
+		"lint": "xo",
+		"autofix": "xo --fix",
+		"test-lib": "cross-env TS_NODE_PROJECT=test/tsconfig.json NODE_OPTIONS=--experimental-vm-modules jest --coverage --detectOpenHandles",
+		"test": "run-s compile lint test-lib",
+		"precommit": "lint-staged",
+		"prepare": "npm run build"
 	},
-	"dependencies": {},
 	"devDependencies": {
-		"@types/express": "^4.17.6",
-		"@types/mocha": "^7.0.2",
-		"@types/node": "^12.0.12",
-		"@types/supertest": "^2.0.9",
-		"@typescript-eslint/parser": "^3.4.0",
-		"cross-env": "^7.0.2",
-		"eslint": "^7.3.1",
-		"eslint-config-prettier": "^6.11.0",
-		"eslint-plugin-prettier": "^3.1.4",
+		"@jest/globals": "^27.3.1",
+		"@types/express": "^4.17.13",
+		"@types/jest": "^27.0.3",
+		"@types/node": "^16.11.10",
+		"@types/sinon": "^10.0.6",
+		"@types/supertest": "^2.0.11",
+		"cross-env": "^7.0.3",
+		"del-cli": "^4.0.1",
 		"express": "^4.17.1",
-		"husky": "^4.2.5",
-		"mocha": "^8.0.1",
-		"prettier": "^2.0.5",
-		"pretty-quick": "^2.0.1",
-		"supertest": "^4.0.2",
-		"ts-node": "^8.10.2",
-		"typescript": "^3.9.5"
+		"husky": "^7.0.4",
+		"jest": "^27.3.1",
+		"lint-staged": "^12.1.2",
+		"npm-run-all": "^4.1.5",
+		"sinon": "^12.0.1",
+		"supertest": "^6.1.6",
+		"ts-jest": "^27.0.7",
+		"ts-node": "^10.4.0",
+		"typescript": "^4.5.2",
+		"xo": "^0.47.0"
+	},
+	"xo": {
+		"prettier": true,
+		"rules": {
+			"@typescript-eslint/no-empty-function": 0,
+			"@typescript-eslint/no-dynamic-delete": 0,
+			"@typescript-eslint/consistent-indexed-object-style": [
+				"error",
+				"index-signature"
+			],
+			"import/no-named-as-default-member": 0,
+			"import/no-cycle": 0
+		}
 	},
 	"prettier": {
 		"semi": false,
@@ -68,5 +85,30 @@
 		"bracketSpacing": true,
 		"trailingComma": "all",
 		"proseWrap": "always"
+	},
+	"jest": {
+		"preset": "ts-jest/presets/default-esm",
+		"globals": {
+			"ts-jest": {
+				"useESM": true
+			}
+		},
+		"testTimeout": 30000,
+		"testMatch": [
+			"**/test/**/*-test.[jt]s?(x)"
+		],
+		"moduleFileExtensions": [
+			"js",
+			"jsx",
+			"json",
+			"ts",
+			"tsx"
+		],
+		"moduleNameMapper": {
+			"^(\\.{1,2}/.*)\\.js$": "$1"
+		}
+	},
+	"lint-staged": {
+		"{source,test}/**/*.ts": "xo --fix"
 	}
 }
